Tweaked the code to optional mode and fix void sizing
I tweaked the code further. However, I left in 'modes' for those that want that behavior, but I changed the default (no mode specified) to block on all edges. I also fixed the need to specify a void that doesn't exactly match the offscreen size. It is still possible to sneak the mouse past the edge, but it is much less likely now, and the simply solution is to just drag the mouse horizontal or vertically (depending on the void orientation) and the mouse will come back (not actually stuck, just lost...). Making the void bigger in the negative space may help too (I didn't try it because I found it very difficult to lose the mouse anymore). I added a makefile, just type 'make' to build (or 'make debug' if you want more verbose output from the executable...).

https://bugs.launchpad.net/ubuntu/+source/libxrandr/+bug/373367/comments/35